# Replace Circe with elasticsearch-x-content to reduce plugin size