--- a/build.gradle
+++ b/build.gradle
@@ -152,11 +152,11 @@ def publishConfig(String projectDescription, boolean isScala = false) {
 
 configure(api4s, List.of(
         scalaProjectConfig,
-        publishConfig("Scala case classes and codecs for the Elastiknn JSON API", true),
+        publishConfig("Scala case classes and JSON codecs for the Elastiknn JSON API", true),
         {
             dependencies {
-                implementation "org.scala-lang:scala-library:${scalaFullVersion}"
                 implementation "org.elasticsearch:elasticsearch-x-content:${esVersion}"
+                implementation "org.scala-lang:scala-library:${scalaFullVersion}"
             }
         }
 ))
@@ -177,19 +177,20 @@ configure(benchmarks, List.of(
             }
 
             dependencies {
-                implementation models
-                implementation lucene
-                implementation plugin
                 implementation clientElastic4s
-                implementation "dev.zio:zio_${scalaShortVersion}:${zioVersion}"
-                implementation "dev.zio:zio-streams_${scalaShortVersion}:${zioVersion}"
-                implementation "dev.zio:zio-logging-slf4j_${scalaShortVersion}:0.4.0"
-                implementation "io.circe:circe-parser_${scalaShortVersion}:${circeVersion}"
-                implementation "io.circe:circe-generic-extras_${scalaShortVersion}:${circeGenericExtrasVersion}"
-                implementation "com.github.scopt:scopt_${scalaShortVersion}:4.0.0-RC2"
+                implementation lucene
+                implementation models
+                implementation plugin
                 implementation "ch.qos.logback:logback-classic:1.2.3"
                 implementation "com.amazonaws:aws-java-sdk-s3:1.11.781"
+                implementation "com.github.scopt:scopt_${scalaShortVersion}:4.0.0-RC2"
                 implementation "com.nrinaudo:kantan.csv-generic_${scalaShortVersion}:0.6.1"
+                implementation "dev.zio:zio_${scalaShortVersion}:${zioVersion}"
+                implementation "dev.zio:zio-logging-slf4j_${scalaShortVersion}:0.4.0"
+                implementation "dev.zio:zio-streams_${scalaShortVersion}:${zioVersion}"
+                implementation "io.circe:circe-generic-extras_${scalaShortVersion}:${circeGenericExtrasVersion}"
+                implementation "io.circe:circe-parser_${scalaShortVersion}:${circeVersion}"
+
                 // Without this you get a warning: WARNING: JAXB is unavailable. Will fallback to SDK implementation which may be
                 // less performant.If you are using Java 9+, you will need to include javax.xml.bind:jaxb-api as a dependency.
                 implementation 'javax.xml:jaxb-api:2.1'
@@ -226,8 +227,8 @@ configure(lucene, List.of(
         {
             dependencies {
                 implementation models
-                implementation "org.apache.lucene:lucene-core:${luceneVersion}"
                 implementation "org.apache.logging.log4j:log4j-api:2.11.1"
+                implementation "org.apache.lucene:lucene-core:${luceneVersion}"
             }
         }))
 
@@ -252,14 +253,12 @@ configure(plugin, List.of(scalaProjectConfig, {
 
     dependencies {
         // TODO: a couple of the deps need both runtime _and_ implementation dependencies to run?
-        implementation models
-        implementation lucene
         implementation api4s
-
+        implementation lucene
+        implementation models
         implementation "com.google.guava:guava:28.1-jre"
-        runtimeOnly 'com.google.guava:failureaccess:1.0.1'
-
         implementation "org.scala-lang:scala-library:${scalaFullVersion}"
+        runtimeOnly "com.google.guava:failureaccess:1.0.1"
         runtimeOnly "org.scala-lang:scala-library:${scalaFullVersion}"
     }
 }))
@@ -279,28 +278,29 @@ configure(testing, List.of(scalaProjectConfig, {
     }
 
     dependencies {
-        implementation models
+        implementation benchmarks
         implementation clientElastic4s
         implementation clientJava
-        implementation plugin
         implementation lucene
-        implementation benchmarks
-        implementation 'com.typesafe:config:1.4.0'
-        implementation "org.scalatest:scalatest_${scalaShortVersion}:3.2.0"
-        implementation "com.typesafe.scala-logging:scala-logging_${scalaShortVersion}:3.9.2"
+        implementation models
+        implementation plugin
         implementation "ch.qos.logback:logback-classic:1.2.3"
-        implementation 'org.apache.commons:commons-math3:3.6.1'
-        implementation "org.apache.lucene:lucene-codecs:${luceneVersion}"
-        implementation "org.apache.lucene:lucene-analyzers-common:${luceneVersion}"
-        implementation "org.elasticsearch:elasticsearch:${esVersion}"
-        implementation "org.elasticsearch.client:elasticsearch-rest-high-level-client:${esVersion}"
-        implementation "com.storm-enroute:scalameter_${scalaShortVersion}:0.19"
-        implementation "org.scalanlp:breeze_${scalaShortVersion}:1.3"
         implementation "com.klibisz.futil:futil_${scalaShortVersion}:0.1.2"
-        implementation "org.scala-lang:scala-library:${scalaFullVersion}"
-        implementation 'org.pegdown:pegdown:1.4.2'
+        implementation "com.storm-enroute:scalameter_${scalaShortVersion}:0.19"
+        implementation 'com.typesafe:config:1.4.0'
+        implementation "com.typesafe.scala-logging:scala-logging_${scalaShortVersion}:3.9.2"
         implementation 'com.vladsch.flexmark:flexmark-all:0.35.10'
-        implementation "io.circe:circe-parser_${scalaShortVersion}:${circeVersion}"
+        implementation "io.circe:circe-literal_${scalaShortVersion}:${circeVersion}"
         implementation "io.circe:circe-generic-extras_${scalaShortVersion}:${circeGenericExtrasVersion}"
+        implementation "io.circe:circe-parser_${scalaShortVersion}:${circeVersion}"
+        implementation 'org.apache.commons:commons-math3:3.6.1'
+        implementation "org.apache.lucene:lucene-analyzers-common:${luceneVersion}"
+        implementation "org.apache.lucene:lucene-codecs:${luceneVersion}"
+        implementation "org.elasticsearch.client:elasticsearch-rest-high-level-client:${esVersion}"
+        implementation "org.elasticsearch:elasticsearch:${esVersion}"
+        implementation 'org.pegdown:pegdown:1.4.2'
+        implementation "org.scala-lang:scala-library:${scalaFullVersion}"
+        implementation "org.scalanlp:breeze_${scalaShortVersion}:1.3"
+        implementation "org.scalatest:scalatest_${scalaShortVersion}:3.2.0"
     }
 }))