--- a/build.gradle
+++ b/build.gradle
@@ -63,8 +63,7 @@ def configure(Project p, List<Closure> configClosures) {
 def scalaProjectConfig = {
     apply plugin: 'scala'
     tasks.withType(ScalaCompile) {
-        // scalaCompileOptions.setAdditionalParameters(List.of("-Xfatal-warnings", "-Ywarn-unused:imports", "-feature"))
-        scalaCompileOptions.setAdditionalParameters(List.of("-Ywarn-unused:imports", "-feature"))
+        scalaCompileOptions.setAdditionalParameters(List.of("-Xfatal-warnings", "-Ywarn-unused:imports", "-feature"))
     }
     task unifiedScaladocs(type: ScalaDoc, description: 'Generate unified scaladocs', group: 'Documentation') {
         Set docProjects = subprojects.findAll { it.plugins.hasPlugin('scala') }