… docs/pages/api.md
… docs/pages/installation.md
… elastiknn-benchmarks/src/main/scala/com/klibisz/elastiknn/api/ElasticsearchCodec.scala
… elastiknn-benchmarks/src/main/scala/com/klibisz/elastiknn/benchmarks/LocalBenchmark.scala
… elastiknn-client-elastic4s/src/main/scala/com/klibisz/elastiknn/client/Elastic4sCompatibility.scala
… elastiknn-client-elastic4s/src/main/scala/com/klibisz/elastiknn/client/ElastiknnClient.scala
… elastiknn-client-elastic4s/src/main/scala/com/klibisz/elastiknn/client/ElastiknnRequests.scala
… elastiknn-plugin/src/main/scala/com/klibisz/elastiknn/ElastiknnException.scala
… elastiknn-plugin/src/main/scala/com/klibisz/elastiknn/query/ElastiknnQuery.scala
… elastiknn-plugin/src/main/scala/com/klibisz/elastiknn/query/KnnQueryBuilder.scala
… elastiknn-plugin/src/main/scala/com/klibisz/elastiknn/utils/CirceUtils.scala
… elastiknn-testing/docker-compose.yml
… elastiknn-testing/src/test/scala/com/klibisz/elastiknn/api/ElasticsearchCodecSuite.scala
… elastiknn-testing/src/test/scala/com/klibisz/elastiknn/api/XContentCodecSuite.scala
… elastiknn-testing/src/test/scala/com/klibisz/elastiknn/mapper/VectorMapperSuite.scala
… elastiknn-testing/src/test/scala/com/klibisz/elastiknn/utils/CirceUtilsSuite.scala
… examples/scala-sbt-client-usage/build.sbt
… version