Re #316 

This removes the Circe dependency and its transitive dependencies in favor of using elasticsearch-x-content, **which gets the plugin down from ~23MB to ~7MB**. The circe dependency remains in the testing and benchmarks projects. 

The new codecs are pretty tedious, but I think this is a worthwhile effort. Copying my reasoning from #316:

1. It will reduce some serde overhead in the plugin itself. Currently there are several places where the plugin code takes an XContentParser, materializes it to a Map, turns that Map into a circe Json ADT, and then decodes the Json to a Scala type. We can avoid a lot of allocations my just going from XContentParser to the Scala type.
2. The current codec is already ~80% manual (i.e., uses the HCursor), and the API is stable. If the codec were largely derived or the API were rapidly evolving, I would see some advantage to learning another library's derivation quirks. However, in this case, we might as well just do the remaining 20% manually and avoid the overhead described in my previous point.

Remaining TODOs:

- [x] Write the remaining codecs. `Vec` could be particularly tricky, as the XContentParser doesn't really let us try to parse a specific vector type and then back track if it turns out to be a different type.
- [x] Write more tests for the XContentCodec failure cases (i.e., test that correct exceptions are thrown).
- [x] Make sure all tests pass.
- [x] Measure performance using the LocalBenchmark.