@alexklibisz 2021-11-22
Benchmark results for the two quick LSH benchmarks look good [compared to the latest results from the Scala 2.13 migration](https://github.com/alexklibisz/elastiknn/pull/317#issuecomment-962514145). 102 QPS -> 107 QPS, 140 QPS -> 153 QPS. The improvement could be noise, but is unlikely to be a regression.

| dataset          | similarity | algorithm | mapping                                                                                                             | query                                                                                 | k   | shards | replicas | parallelQueries | esNodes | esCoresPerNode | esMemoryGb | warmupQueries | minWarmupRounds | maxWarmupRounds | recall     | queriesPerSecond | durationMillis |
| ---------------- | ---------- | --------- | ------------------------------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------------------- | --- | ------ | -------- | --------------- | ------- | -------------- | ---------- | ------------- | --------------- | --------------- | ---------- | ---------------- | -------------- |
| AnnbFashionMnist | "l2"       | LSH       | {"elastiknn":{"L":75,"dims":784,"k":4,"model":"lsh","similarity":"l2","w":7},"type":"elastiknn_dense_float_vector"} | {"candidates":2000,"field":"vec","model":"lsh","probes":3,"similarity":"l2","vec":{}} | 100 | 1      | 0        | 1               | 1       | 1              | 4          | 200           | 10              | 10              | 0.856039   | 107.52688        | 93728          |
| AnnbFashionMnist | "l2"       | LSH       | {"elastiknn":{"L":75,"dims":784,"k":4,"model":"lsh","similarity":"l2","w":7},"type":"elastiknn_dense_float_vector"} | {"candidates":1000,"field":"vec","model":"lsh","probes":0,"similarity":"l2","vec":{}} | 100 | 1      | 0        | 1               | 1       | 1              | 4          | 200           | 10              | 10              | 0.53446233 | 153.84616        | 65895          |


VisualVM hotspots also look reasonable. Biggest JSON parsing hotspot is < 1% of CPU time.

![image](https://user-images.githubusercontent.com/8015228/142803379-75eb8801-3b6b-4980-9794-cbf59ce06859.png)