@fabriziofortino 2021-11-22
@alexklibisz The code looks okay to me.

Perhaps is not related to these changes but when I run the entire test suite locally, 3 tests fail:

```
- freeze index without elastiknn setting *** FAILED *** (2 seconds, 392 milliseconds)
  java.lang.RuntimeException: parse_exception request body is required
  at com.sksamuel.elastic4s.ElasticError.$anonfun$asException$1(ElasticError.scala:18)
  at scala.Option.fold(Option.scala:263)
  at com.sksamuel.elastic4s.ElasticError.asException(ElasticError.scala:18)
  at com.klibisz.elastiknn.client.ElastiknnClient$.com$klibisz$elastiknn$client$ElastiknnClient$$checkResponse(ElastiknnClient.scala:174)
```

```
RecallSuite:
- 9ab5b94e-22b0-4dc4-add5-56b114f9eeca JaccardLsh(1024,300,2)         JaccardLsh(vec,800,Empty())                        ~= 0,810000 *** FAILED *** (13 seconds, 206 milliseconds)
  Explicit query recall 0.7998 was not 0.81 plus or minus 0.01 (RecallSuite.scala:269)
  org.scalatest.exceptions.TestFailedException:
  at org.scalatest.matchers.MatchersHelper$.indicateFailure(MatchersHelper.scala:339)
  at org.scalatest.matchers.should.Matchers$AnyShouldWrapper.shouldBe(Matchers.scala:7100)
  at com.klibisz.elastiknn.query.RecallSuite.$anonfun$new$26(RecallSuite.scala:269)
```

```
ClusterSpec:
- returns health (138 milliseconds)
- installed the plugin (24 milliseconds)
- started four nodes *** FAILED *** (99 milliseconds)
  List(CatNodesResponse(jXYH,127.0.0.1,69423,12.0,17.5m,18.0,100.0,9300,5.24,0.0,0.0,cdfhimrsw,*,runTask-0)) had length 1 instead of expected length 2 (ClusterSpec.scala:35)
  org.scalatest.exceptions.TestFailedException:
  at org.scalatest.matchers.MatchersHelper$.indicateFailure(MatchersHelper.scala:339)
  at org.scalatest.matchers.should.Matchers$ResultOfHaveWordForExtent.length(Matchers.scala:2698)
  at com.klibisz.elastiknn.ClusterSpec.$anonfun$new$6(ClusterSpec.scala:35)
```