@alexklibisz 2021-11-23
>   java.lang.RuntimeException: parse_exception request body is required

I did see this one yesterday. This came up when I was testing against a cluster running via gradle. It goes away when I test against the cluster running docker.

I didn't see the other two. What's your testing environment? Running via gradle or via docker or something else?