@fabriziofortino 2021-11-23
macOS Big Sur v 11.6

I start the ES cluster with

```
task jvm:run:gradle
```

and run the Gradle test suite with

```
task jvm:test
```