@alexklibisz 2021-11-23
Thanks @fabriziofortino . The first and third failures are to be expected with a single-node ES deployment, which is the case when running from Gradle. The second one is strange, but given they are passing reliably on GH, I feel comfortable proceeding.

I added some tests and better error handling in last commit. I'll add maybe ten more tests for error handling and call this done. Thanks for your help reviewing.